refactor(skills): move the verified server-skills read path into the SDK
`server-skills.ts` claims to be the ONE module between every MCPJam surface and
a `skills/*` call — "the SEP's integrity rules are only rules if there is no
second way in" — and its header names the CLI as a consumer. That was
aspiration: living in `mcpjam-inspector/server`, it was unreachable from `cli/`
and `mcp/`, so any skills surface those packages grow would have had to
re-decide which server behaviours are refusals. A second implementation of the
integrity rules is the exact thing the module exists to prevent.

The move is byte-faithful except three mechanical changes:

- imports become relative (`./mcp-client-manager/index.js` for the integrity
  primitives, `./operations.js` for `listAllServerSkills`);
- `Buffer.byteLength` becomes `TextEncoder`, since the SDK's browser entry bans
  node built-ins — the same reason `skills-integrity.ts` already does this;
- the module-level `logger` import, which is precisely what could not travel
  (the CLI and the worker have their own), becomes an optional injected
  `{ logger }` on `probeServerSkillMissing` — its one inconclusive-probe
  diagnostic. No production caller exists today, so nothing had to be threaded.

Placement is top-level rather than `mcp-client-manager/`: the module needs
`listAllServerSkills` from `operations.ts`, which imports from
`mcp-client-manager/`, so living there would close an import cycle.

`server/utils/server-skills.ts` stays as a re-export, so no inspector call site
changed — routes, chat tools and the capture coordinator keep their import path
and this diff stays a diff of imports rather than of behaviour.

The 28 cases move with it and pass unmodified apart from the byte-length helper
(git records the test as a rename). Consumers: 564 inspector tests green; SDK
skills suites 113 green including the browser-purity guard; server typecheck
shows zero errors mentioning the module.
